RFC: Add NoSemicolonAfterShortEchoTagFixer (see #5)
Description
-----------

I've noticed this (html5 template) code style all over the templates in the `contao/*` ecosystem but there is no fixer for this kind of code style.

As I want to use this code style at our daily work, I've come up with a fixer for that.

WDYT? Would this also be useful for Contao or should I just make my own ECS package for that?

Commits
-------

b1ec1253 Add NoSemicolonAfterSingleEchoInstructionFixer
21dfe232 Only apply to T_OPEN_TAG_WITH_ECHO - renmaed fixer
01fa0cf7 CS